% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestochile-package.R
\docType{data}
\name{capitulos}
\alias{capitulos}
\title{Capitulos (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada capitulo, es
decir para el Senado y Camara de Diputados que pertenecen al Congreso
Nacional, etc.}
\format{
Un tibble de 246 filas y 9 columnas
}
\usage{
capitulos
}
\description{
Capitulos (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada capitulo, es
decir para el Senado y Camara de Diputados que pertenecen al Congreso
Nacional, etc.
}
\examples{
\dontrun{
capitulos \%>\%
    left_join(partidas \%>\% select(nombre_partida, id_partida)) \%>\%
    select(nombre_partida, everything())
}
}
\references{
\url{http://presupuesto.bcn.cl/presupuesto/api}
}
\author{
Direccion de Presupuestos (DIPRES)
}
\keyword{data}
