% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos_receptor.R
\name{descargar_datos_receptor}
\alias{descargar_datos_receptor}
\title{Download JSON files from 'presupuestoabierto.gob.cl' by recipient identification number}
\usage{
descargar_datos_receptor(ruts, years, val = FALSE)
}
\arguments{
\item{ruts}{A vector with the 'rut' of each recipient to download.}

\item{years}{A vector with the years to download.}

\item{val}{Logical. Delete variables that contain only NA? FALSE by default.}
}
\value{
Return a data set with all the information of recipients of public resources per year.
}
\description{
This function has been designed to download and consolidate JSON files from the
webpage 'presupuestoabierto.gob.cl'. The output is a data set with all the information of
recipients of public resources per year.
}
\examples{
\dontrun{
descargar_datos_receptor(c("70005600-7","81591900-9"),c(2016,2017,2018,2019),val=TRUE)
}
}
\seealso{
\link[jsonlite]{fromJSON}

\link[dplyr]{select}
}
\author{
Hugo Mansilla
}
