% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos_region.R
\name{descargar_datos_region}
\alias{descargar_datos_region}
\title{Download JSON files from 'presupuestoabierto.gob.cl' by region}
\usage{
descargar_datos_region(region, years)
}
\arguments{
\item{region}{A vector with the number of each region to download.}

\item{years}{A vector with the years to download.}
}
\value{
Return a data set with all the information of public resources by region per year.
}
\description{
This function has been designed to download and consolidate JSON files from the
webpage 'presupuestoabierto.gob.cl'. The output is a data set with all the information
of public resources by region of Chile per year.
}
\examples{
\dontrun{
descargar_datos_region(region=12,c(2016,2017,2018,2019,2020))
}
}
\seealso{
\link[jsonlite]{fromJSON}

\link[dplyr]{select}
}
\author{
Hugo Mansilla
}
