% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestochile-package.R
\docType{data}
\name{partidas}
\alias{partidas}
\title{Partidas (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada partida, es
decir para el Poder Judicial, Contraloria, etc.}
\format{
Un tibble de 246 filas y 9 columnas
}
\usage{
partidas
}
\description{
Partidas (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada partida, es
decir para el Poder Judicial, Contraloria, etc.
}
\references{
\url{http://presupuesto.bcn.cl/presupuesto/api}
}
\author{
Direccion de Presupuestos (DIPRES)
}
\keyword{data}
