% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestochile.r
\docType{package}
\name{presupuestochile}
\alias{presupuestochile}
\alias{presupuestochile-package}
\title{presupuestochile: Download JSON files from 'presupuestoabierto.gob.cl'.}
\value{
Return a data set with all the information of recipients of public resources per year.
}
\description{
This package has been designed to download and consolidate JSON files from the
webpage 'presupuestoabierto.gob.cl'.
}
\examples{
\dontrun{
descargar_datos_receptor(c("70005600-7","81591900-9"),c(2016,2017,2018,2019),val=TRUE)
}
\dontrun{
descargar_datos_region(region=12,c(2016,2017,2018,2019,2020))
}
}
\author{
\strong{Maintainer}: Hugo Mansilla \email{hugomansilla@gmail.com}

}
