% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestochile-package.R
\docType{data}
\name{presupuestos}
\alias{presupuestos}
\title{Presupuestos (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion por anio.}
\format{
Un tibble de 9 filas y 8 columnas
}
\usage{
presupuestos
}
\description{
Presupuestos (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion por anio.
}
\references{
\url{http://presupuesto.bcn.cl/presupuesto/api}
}
\author{
Direccion de Presupuestos (DIPRES)
}
\keyword{data}
