% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestochile-package.R
\docType{data}
\name{programas}
\alias{programas}
\title{Programas (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada programa, es
decir para la Academia Judicial y la Corporacion Administrativa del
Poder Judicial, etc.}
\format{
Un tibble de 2874 filas y 17 columnas
}
\usage{
programas
}
\description{
Programas (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada programa, es
decir para la Academia Judicial y la Corporacion Administrativa del
Poder Judicial, etc.
}
\examples{
\dontrun{
programas \%>\%
  left_join(capitulos \%>\% select(id_capitulo, id_partida)) \%>\%
  left_join(partidas \%>\% select(nombre_partida, id_partida)) \%>\%
  select(nombre_partida, everything())
}
}
\references{
\url{http://presupuesto.bcn.cl/presupuesto/api}
}
\author{
Direccion de Presupuestos (DIPRES)
}
\keyword{data}
