% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestochile-package.R
\docType{data}
\name{subtitulos}
\alias{subtitulos}
\title{Subtitulos (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada subtitulo, es
decir para el aporte fiscal, gastos en personal, saldo de caja de cada programa
del Poder Judicial, etc.}
\format{
Un tibble de 33852 filas y 13 columnas
}
\usage{
subtitulos
}
\description{
Subtitulos (2012-2020)
Contiene el valor asignado al Presupuesto de la Nacion para cada subtitulo, es
decir para el aporte fiscal, gastos en personal, saldo de caja de cada programa
del Poder Judicial, etc.
}
\examples{
\dontrun{
subtitulos \%>\%
  left_join(programas \%>\% select(id_programa, id_capitulo)) \%>\%
  left_join(capitulos \%>\% select(id_capitulo, id_partida)) \%>\%
  left_join(partidas \%>\% select(nombre_partida, id_partida)) \%>\%
  select(nombre_partida, everything())
}
}
\references{
\url{http://presupuesto.bcn.cl/presupuesto/api}
}
\author{
Direccion de Presupuestos (DIPRES)
}
\keyword{data}
